# swagger-docs/openapi.json
{"openapi":"3.0.0","info":{"title":"Moteur API","version":"1.0.0","description":"REST API for the Moteur content engine"},"paths":{"/generate-entry":{"post":{"summary":"Generate an entry using AI based on model schema","tags":["AI"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"prompt":{"type":"string","description":"User-provided description of the content to generate","example":"Create a product entry for an eco-friendly water bottle"},"projectId":{"type":"string","description":"ID of the current project"},"modelId":{"type":"string","description":"ID of the model schema to use"},"locale":{"type":"string","description":"Locale to use for multilingual fields","example":"en","default":"en"}},"required":["prompt","projectId","modelId"]}}}},"responses":{"200":{"description":"Entry successfully generated","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"entry":{"type":"object","properties":{"data":{"type":"object","description":"Generated entry data, keyed by field","additionalProperties":true}}}}}}}},"400":{"description":"Bad Request (missing or invalid parameters)","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"},"issues":{"type":"array","items":{"type":"object","properties":{"type":{"type":"string"},"code":{"type":"string"},"message":{"type":"string"},"path":{"type":"string"},"hint":{"type":"string"}}}}}}}}},"500":{"description":"Internal Server Error (AI failure)","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/generate-fields":{"post":{"summary":"Generate field definitions using AI","tags":["AI"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"prompt":{"type":"string","description":"User-provided description of the model","example":"Create fields for a blog post"},"currentFields":{"type":"object","description":"Current field definitions (keyed by field name)","additionalProperties":{"type":"object","properties":{"type":{"type":"string"},"label":{"type":"string"}},"required":["type"]}}},"required":["prompt","currentFields"]}}}},"responses":{"200":{"description":"Field definitions successfully generated","content":{"application/json":{"schema":{"type":"object","properties":{"fields":{"type":"object","additionalProperties":{"type":"object","properties":{"type":{"type":"string"},"label":{"type":"string"},"meta":{"type":"object"},"schema":{"type":"string"}},"required":["type"]}}}}}}},"400":{"description":"Bad Request (invalid input)","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}},"500":{"description":"Internal Server Error (AI failure)","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/auth/login":{"post":{"summary":"Login and receive JWT token","tags":["Auth"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginInput"}}}},"responses":{"200":{"description":"Successful login","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}}}}}},"400":{"description":"Missing fields or invalid format","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/auth/providers":{"get":{"summary":"List enabled auth providers","description":"Returns a list of OAuth providers (e.g., GitHub, Google) that are currently enabled based on environment config.","tags":["Auth"],"responses":{"200":{"description":"Enabled providers","content":{"application/json":{"schema":{"type":"object","properties":{"providers":{"type":"array","items":{"$ref":"#/components/schemas/AuthProvider"}}}}}}}}}},"/auth/refresh":{"post":{"summary":"Refresh the JWT","tags":["Auth"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"New JWT token","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string"}}}}}},"401":{"description":"Unauthorized or invalid user","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/auth/me":{"get":{"summary":"Get current authenticated user","tags":["Auth"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Current user","content":{"application/json":{"schema":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"}}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/projects":{"get":{"summary":"List all accessible projects","tags":["Projects"],"responses":{"200":{"description":"List of projects","content":{"application/json":{"schema":{"type":"object","properties":{"projects":{"type":"array","items":{"$ref":"#/components/schemas/Project"}}}}}}}}},"post":{"summary":"Create a new project","tags":["Projects"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"responses":{"201":{"description":"Project successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewProjectInput"}}}},"400":{"description":"Validation failed or bad input","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"},"validation":{"type":"array","items":{"type":"object","properties":{"path":{"type":"string"},"message":{"type":"string"}}}}}}}}}}}},"/projects/{projectId}":{"get":{"summary":"Get a single project by ID","tags":["Projects"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The requested project","content":{"application/json":{"schema":{"type":"object","properties":{"project":{"$ref":"#/components/schemas/Project"}}}}}},"404":{"description":"Project not found","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}},"patch":{"summary":"Update a project","tags":["Projects"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectInput"}}}},"responses":{"200":{"description":"Project updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"},"validation":{"type":"array","items":{"type":"object","properties":{"path":{"type":"string"},"message":{"type":"string"}}}}}}}}}}},"delete":{"summary":"Delete a project","tags":["Projects"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Project deleted successfully"},"404":{"description":"Project not found","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/projects/{projectId}/users":{"get":{"summary":"List users with access to a project","tags":["Users"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"List of users for the project","content":{"application/json":{"schema":{"type":"object","properties":{"users":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"400":{"description":"Missing project ID","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}},"500":{"description":"Internal server error"}}}},"/projects/{projectId}/presence/form-state/{screenId}":{"delete":{"summary":"Clear form state for a specific screen","tags":["Presence"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"screenId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Form state cleared","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}}}}}},"400":{"description":"Missing screenId","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/projects/{projectId}/presence/debug":{"get":{"summary":"Inspect presence and form state in a project","tags":["Presence"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Active presence and form state","content":{"application/json":{"schema":{"type":"object","properties":{"presence":{"type":"array","items":{"$ref":"#/components/schemas/Presence"}},"formState":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"400":{"description":"Missing projectId","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/projects/{projectId}/models":{"get":{"summary":"List models in a project","tags":["Models"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"List of models in the given project","content":{"application/json":{"schema":{"type":"object","properties":{"models":{"type":"array","items":{"$ref":"#/components/schemas/Model"}}}}}}}}},"post":{"summary":"Create a new model in a project","tags":["Models"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewModelInput"}}}},"responses":{"201":{"description":"Model successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Model"}}}},"400":{"description":"Validation failed or invalid model","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"},"validation":{"type":"array","items":{"type":"object","properties":{"path":{"type":"string"},"message":{"type":"string"}}}}}}}}}}}},"/projects/{projectId}/models/{modelId}":{"get":{"summary":"Get a single model from a project","tags":["Models"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"modelId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Returns the model schema","content":{"application/json":{"schema":{"type":"object","properties":{"model":{"$ref":"#/components/schemas/Model"}}}}}},"404":{"description":"Model not found","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}},"patch":{"summary":"Update a model in a project","tags":["Models"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"modelId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateModelInput"}}}},"responses":{"200":{"description":"Model updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Model"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"},"validation":{"type":"array","items":{"type":"object","properties":{"path":{"type":"string"},"message":{"type":"string"}}}}}}}}}}},"delete":{"summary":"Delete a model from a project","tags":["Models"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"modelId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Model deleted successfully"},"404":{"description":"Model not found","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/projects/{projectId}/models/{modelId}/entries":{"get":{"summary":"List entries for a model","tags":["Entries"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"modelId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"List of entries","content":{"application/json":{"schema":{"type":"object","properties":{"entries":{"type":"array","items":{"type":"object"}}}}}}},"400":{"description":"Missing parameters","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}},"post":{"summary":"Create a new entry","tags":["Entries"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"modelId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"responses":{"201":{"description":"Entry created","content":{"application/json":{"schema":{"type":"object"}}}},"400":{"description":"Validation failed or invalid input","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationResult"}}}}}}},"/projects/{projectId}/models/{modelId}/entries/{entryId}":{"get":{"summary":"Get a single entry","tags":["Entries"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"modelId","in":"path","required":true,"schema":{"type":"string"}},{"name":"entryId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Single entry found","content":{"application/json":{"schema":{"type":"object","properties":{"entry":{"type":"object"}}}}}},"404":{"description":"Entry not found","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}},"patch":{"summary":"Update an entry","tags":["Entries"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"modelId","in":"path","required":true,"schema":{"type":"string"}},{"name":"entryId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"responses":{"200":{"description":"Entry updated","content":{"application/json":{"schema":{"type":"object"}}}},"400":{"description":"Validation failed or bad input","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationResult"}}}}}},"delete":{"summary":"Delete an entry","tags":["Entries"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"modelId","in":"path","required":true,"schema":{"type":"string"}},{"name":"entryId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Entry deleted successfully"},"404":{"description":"Entry not found","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}}},"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}},"schemas":{"User":{"type":"object","required":["id","email"],"properties":{"id":{"type":"string"},"email":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"projects":{"type":"array","items":{"type":"string"}},"isActive":{"type":"boolean"}}},"Project":{"type":"object","required":["id","label"],"properties":{"id":{"type":"string"},"label":{"type":"string"},"description":{"type":"string"},"plugins":{"type":"array","items":{"type":"string"}}}},"Model":{"type":"object","required":["id","label","fields"],"properties":{"id":{"type":"string"},"label":{"type":"string"},"description":{"type":"string"},"fields":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Field"}},"meta":{"type":"object","additionalProperties":true}}},"Field":{"type":"object","required":["type"],"properties":{"type":{"type":"string"},"label":{"type":"string"},"required":{"type":"boolean"},"options":{"type":"object","additionalProperties":true},"meta":{"type":"object","additionalProperties":true}}},"ValidationResult":{"type":"object","required":["valid","errors"],"properties":{"valid":{"type":"boolean","example":false},"errors":{"type":"array","items":{"type":"object","required":["field","message"],"properties":{"field":{"type":"string"},"message":{"type":"string"}}}}}},"Presence":{"type":"object","properties":{"userId":{"type":"string"},"name":{"type":"string"},"avatarUrl":{"type":"string","nullable":true},"projectId":{"type":"string"},"screenId":{"type":"string","nullable":true},"entryId":{"type":"string","nullable":true},"fieldPath":{"type":"string","nullable":true},"typing":{"type":"boolean","nullable":true},"textPreview":{"type":"string","nullable":true},"cursor":{"type":"object","nullable":true,"properties":{"x":{"type":"number"},"y":{"type":"number"}}},"updatedAt":{"type":"number"}},"required":["userId","name","projectId","updatedAt"]},"LoginInput":{"type":"object","required":["username","password"],"properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AuthProvider":{"type":"object","required":["id","label"],"properties":{"id":{"type":"string","description":"Unique ID of the provider (e.g., github, google)"},"label":{"type":"string","description":"Human-readable label (e.g., GitHub)"}}},"NewProjectInput":{"type":"object","required":["id","label"],"properties":{"id":{"type":"string"},"label":{"type":"string"},"description":{"type":"string"},"locale":{"type":"string"},"modules":{"type":"array","items":{"type":"string"}},"plugins":{"type":"array","items":{"type":"string"}}}}}},"security":[{"bearerAuth":[]}]}